Added StatusService to convert error messages to text/plain

+++ .travis/oracle/install.sh
@@ -1,11 +1,14 @@
 #!/bin/sh -e
 
-ORACLE_RPM=`basename $ORACLE_FILE .zip`
+[ -n "$ORACLE_FILE" ] || { echo "Missing ORACLE_FILE environment variable!"; exit 1; }
+[ -n "$ORACLE_HOME" ] || { echo "Missing ORACLE_HOME environment variable!"; exit 1; }
+
+ORACLE_RPM="$(basename $ORACLE_FILE .zip)"
 
 cd "$(dirname "$(readlink -f "$0")")"
 
 sudo apt-get -qq update
-sudo apt-get --no-install-recommends -qq install alien bc libaio1 unzip
+sudo apt-get --no-install-recommends -qq install bc libaio1 rpm unzip
 
 df -B1 /dev/shm | awk 'END { if ($1 != "shmfs" && $1 != "tmpfs" || $2 < 2147483648) exit 1 }' ||
   ( sudo rm -r /dev/shm && sudo mkdir /dev/shm && sudo mount -t tmpfs shmfs -o size=2G /dev/shm )
@@ -15,21 +18,15 @@ test -f /sbin/chkconfig ||
 
 test -d /var/lock/subsys || sudo mkdir /var/lock/subsys
 
-unzip -j "$ORACLE_FILE" "*/$ORACLE_RPM"
-sudo dpkg --install `sudo alien --scripts --to-deb "$ORACLE_RPM" | cut -d' ' -f1`
+unzip -j "$(basename $ORACLE_FILE)" "*/$ORACLE_RPM"
+sudo rpm --install --nodeps --nopre "$ORACLE_RPM"
 
 echo 'OS_AUTHENT_PREFIX=""' | sudo tee -a "$ORACLE_HOME/config/scripts/init.ora" > /dev/null
 sudo usermod -aG dba $USER
 
 ( echo ; echo ; echo travis ; echo travis ; echo n ) | sudo AWK='/usr/bin/awk' /etc/init.d/oracle-xe configure
 
-"$ORACLE_HOME/bin/sqlplus" -L / AS SYSDBA <<SQL
-SHOW PARAMETER TARGET;
-alter system set memory_max_target = 412m scope = spfile;
-alter system set memory_target = 412m scope = spfile;
-ALTER SYSTEM SET SGA_TARGET = 0 scope = spfile;
-ALTER SYSTEM SET PGA_AGGREGATE_TARGET = 0 scope = spfile;
-shutdown immediate
-startup
-exit
+"$ORACLE_HOME/bin/sqlplus" -L -S / AS SYSDBA <<SQL
+CREATE USER $USER IDENTIFIED EXTERNALLY;
+GRANT CONNECT, RESOURCE TO $USER;
 SQL
